requirements.txt
phidata 
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
python-multipart
groq
openai
pypdf
pandas
llama-parse
llama-index